+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Wrap all requester message text to automatically fit
 - Status bar no longer displays an unwanted trailing character
 - Edit menu items are disabled unless the chat or image input text field has the keyboard focus
+- Improve speech error handling to fix crashes
 
 ## 3.0.0 (2026-07-01)